smartcommhub-backend/requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.30
sqlalchemy~=2.0
alembic~=1.13
psycopg2-binary~=2.9.9
pydantic~=2.6
pydantic-settings~=2.4
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
cryptography>=41,<43
